Show virtualenv name in bash prompt
This is needed because virtualenv's activate script prepends to $PS1,
and then we overwrite that name because we emit a \r to move to start of
the line (for a good reason, too).

Also fix bug where we'd show $debian_choot and then immediatelly
override it in the same fashion.