Avoid beeping in screen in Linux virtual consoles
The xterm title sequence causes garbage on screen and a beep when
emitted on a Linux virtual console, which doesn't support it.  Add a
guard clause (check if $DISPLAY is set) before attempting to use it.
I cannot use $TERM because it might be "screen" in both virtual consoles
and xterms.

I'll have to rethink this when Wayland arrives.

The diff is gnarly because I decided to split the decision and
action into separate parts.  There's a side effect that I now set a
DEBUG trap to run update_cmdstart even if $MC_SID is set.  This didn't
appear to work inside Midnight Commander in my brief tests but oh well.