I think I figured out my branch-diff problem
When somebody rebases on origin/master, and my local master is outdated,
the old branch-diff would show some origin/master changes in addition to the
branch changes I wanted to see.

I think it's unlikely that I'll forget to push master and then switch to
review a branch that has commits from my new master that are not yet in
origin/master, so this new alias should be good.

I wonder if I can ask git to give me the newer of two refs, assuming
they're linear.