Avoid variable clutter in bashrc.prompt
I'm hoping to fix the very weird problem I have where on a fresh new
machine

    git clone https://gedmin.as/dotfiles
    . dotfiles/install.sh

leaves me with a broken prompt.

Specifically, `colors` does not get defined as the array of colors I
expect, but contains a value like "/home/liveuser/.dircolors", and the
expansion in _prompt_color somehow tries to evaluate the value of $green
instead of using 'green' as the lookup key.  I don't understand bash.