Fix the initial . dotfiles/install.sh
Looks like you cannot declare global associative arrays if you're inside
a function and source another file.

The unset is for when I mess up while experimenting with things like
colors=(), because bash says that it cannot convert an indexed array to
an associatie array, but it's happy to undefine and redefine it.