mpv: disable wheel events
Wheel up/down jumps around in time, which is annoying when it happens by
accident (and it _always_ happens by accident).

Wheel left/right changes volume, but is wayyy too sensitive so I
basically can set volume to 0 or to 120% by accidentally two-finger
scrolling left/right on the touchpad, neither of which is useful.