Remove decoy
This was supposed to be renamed to 'vimrc' when I renamed all the other
dotfiles in 84ec4fa512fd56ea10e1777a5eb05d5090f32ee4, but I forgot.
It was then supposed to be replaced with a regular file in
8c408c56f78df5a14524a8218b7c185d961e4b29, but since I forgot to rename,
I added a new file instead.

Anyway, this is just a decoy, not installed by install.sh, and not used
in any other way.