Work work
I like to

  git clone https://github.com/ProgrammersOfVilnius/somerepo

because I can just paste the URL from my browser, and also because then
I can use my git open-url alias to open the web page in the browser.

Some of our repos are public and this works fine.

Some of our repos are private and this fails.  So let's do the ssh
syntax for git pulls too.