Make sure we don't lose VTE integration
There are escape sequences that identify the location of the prompt, and
the exit status of the previous command.

I wonder if it works?  My $PROMPT_COMMAND might be overriding $_

I wonder how to test that it works?  Supposedly Ptyxis does something
with this (notifications for long-running jobs?), but I've never seen
what it's supposed to look like.